+++ Editor/NRenderTexturePreview.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Reflection;
 using UnityEditor;
 using UnityEngine;
 using Object = UnityEngine.Object;
@@ -12,19 +11,19 @@ namespace Vertx
 		new void OnEnable()
 		{
 			//When this inspector is created, also create the built-in inspector
-			defaultEditor = CreateEditor(targets, Type.GetType("UnityEditor.RenderTextureEditor, UnityEditor"));
+			if(defaultEditor == null)
+				defaultEditor = CreateEditor(targets, Type.GetType("UnityEditor.RenderTextureEditor, UnityEditor"));
 			base.OnEnable();
 		}
 
 		protected override void OnDisable()
 		{
 			base.OnDisable();
-			//When OnDisable is called, the default editor we created should be destroyed to avoid memory leakage.
-			//Also, make sure to call any required methods like OnDisable
-			MethodInfo disableMethod = defaultEditor.GetType().GetMethod("OnDisable", BindingFlags.Instance | BindingFlags.NonPublic | BindingFlags.Public);
-			if (disableMethod != null)
-				disableMethod.Invoke(defaultEditor, null);
-			DestroyImmediate(defaultEditor);
+			if (defaultEditor != null)
+			{
+				DestroyImmediate(defaultEditor);
+				defaultEditor = null;
+			}
 		}
 
 		public override void OnInspectorGUI()